+++ llm/turbine_llm/layers/kv_cache.py
@@ -16,6 +16,11 @@ import math
 import torch
 
 
+__all__ = [
+    "PagedKVCache",
+]
+
+
 class PagedKVCache:
     """Implementation of a KV cache on top of a 'page table'.
